aocli: Move pipe reader into its own thread
Opening a named pipe for read blocks until something else opens it for
write. That's great, except if we have a syntax error in our day solver
we'll never open the pipe, and lock up aocli forever. Now we put the
pipe reader into a thread and use a queue for communicating.