24d09 p2 optimization
The vast majority of our time is spent searching for the leftmost free space
that will fit a given file. Because the input is given as single-character
sizes, we know there are only 10 distinct sizes for free space. So we store free
space in 10 min-heaps, one for each size, prioritized on block id (index into
the filesystem). This speeds up part2 from ~1.8s to ~15ms.