clean up graph_dyn types
All of my path-finding helpers just need a dict-like object, mapping
nodes to the set of their neighbors/edges. Introduced an Edges alias for
Mapping[T, set[T]], which all the helpers now require, and updated
DiGraph to conform to that API using the Mapping ABC.

Now it's easy to use either a DiGraph subclass if dynamic neighbor
generation makes sense, or just to build an edges dict directly.

24d23 cleanup